changed posix data types to ones recognized by both linux and windows compilers

+++ include/illixr/record_logger.hpp
@@ -73,7 +73,7 @@ public:
         return columns_[column].second;
     }
 
-    [[nodiscard]] unsigned get_columns() const {
+    [[nodiscard]] size_t get_columns() const {
         return columns_.size();
     }
